# src/main/resources/static/swagger_my05030.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "🍊Discodeit API 문서",
    "description": "🍊Discodeit 프로젝트의 Swagger API 문서입니다."
  },
  "servers": [
    {
      "url": "http://localhost:4444",
      "description": "로컬 서버"
    }
  ],
...
  "paths": {
    "/api/users": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "전체 User 목록 조회",
        "operationId": "findAll",
        "responses": {
          "200": {
            "description": "User 목록 조회 성공",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "User"
        ],
        "summary": "User 등록",
        "operationId": "create",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "userCreateRequest": {
                    "$ref": "#/components/schemas/UserCreateRequest"
                  },
                  "profile": {
                    "type": "string",
                    "format": "binary",
                    "description": "User 프로필 이미지"
                  }
                },
                "required": [
                  "userCreateRequest"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User가 성공적으로 생성됨",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "400": {
            "description": "같은 email 또는 username를 사용하는 User가 이미 존재함",
            "content": {
              "*/*": {
                "example": "User with email {email} already exists"
              }
            }
          }
        }
      }
    },
...
    "/api/users/{userId}": {
      "delete": {
        "tags": [
          "User"
        ],
        "summary": "User 삭제",
        "operationId": "delete",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "삭제할 User ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "User가 성공적으로 삭제됨",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "User를 찾을 수 없음",
            "content": {
              "*/*": {
                "example": "User with id {userId} not found"
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "User"
        ],
        "summary": "User 정보 수정 (프로필 이미지 포함)",
        "operationId": "update",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "수정할 User ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "userUpdateRequest": {
                    "$ref": "#/components/schemas/Dto_UserUpdate"
                  },
                  "profile": {
                    "type": "string",
                    "format": "binary",
                    "description": "새로운 프로필 이미지 파일 (선택 사항)"
                  }
                },
                "required": [
                  "userUpdateRequest"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User 정보가 성공적으로 수정됨",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "404": {
            "description": "User를 찾을 수 없음",
            "content": {
              "*/*": {
                "example": "User with id {userId} not found"
              }
            }
          },
          "400": {
            "description": "같은 email 또는 username를 사용하는 User가 이미 존재함",
            "content": {
              "*/*": {
                "example": "user with email {newEmail} already exists"
              }
            }
          }
        }
      }
    },
    "/api/users/{userId}/userStatus": {
      "patch": {
        "tags": [
          "User"
        ],
        "summary": "User 온라인 상태 업데이트",
        "operationId": "updateUserStatus",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "상태를 변경할 User ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Dto_UserStatusUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "User 온라인 상태가 성공적으로 업데이트됨",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "404": {
            "description": "해당 User의 UserStatus를 찾을 수 없음",
            "content": {
              "*/*": {
                "example": "UserStatus with userId {userId} not found"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "UserCreateRequest": {
        "type": "object",
        "description": "User 생성 정보 (JSON)",
        "properties": {
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password",
          "username"
        ]
      },
      "BinaryContentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "fileName": {
            "type": "string"
          },
          "size": {
            "type": "integer",
            "format": "int64"
          },
          "contentType": {
            "type": "string"
          }
        }
      },
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "profile": {
            "$ref": "#/components/schemas/BinaryContentDto"
          },
          "online": {
            "type": "boolean"
          }
        }
      },
...
      "Dto_UserUpdate": {
        "type": "object",
        "description": "User 업데이트 정보 (JSON)",
        "properties": {
          "newUsername": {
            "type": "string"
          },
          "newEmail": {
            "type": "string"
          },
          "newPassword": {
            "type": "string"
          }
        },
        "required": [
          "newEmail",
          "newPassword",
          "newUsername"
        ]
      },
      "Dto_UserStatusUpdate": {
        "type": "object",
        "properties": {
          "newLastActiveAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "newLastActiveAt"
        ]
      },
...
    }
  }
}